take out form Components from functions Components to prevent inputs loosing focus; fix all issues with login and sigup
